refactor(commands)!: change the way of disabling commands
BREAKING CHANGE: not it does not require calling .setup()
